benchmarks/runner.py
#!/usr/bin/env python3
"""
Benchmark Runner for Trading System Integration

Runs performance benchmarks and generates reports.
"""

import asyncio
import time
import psutil
import json
from datetime import datetime
from pathlib import Path

async def run_benchmarks():
    """Run all benchmark tests"""
    results = {
        "timestamp": datetime.now().isoformat(),
        "system_info": {
            "cpu_count": psutil.cpu_count(),
            "memory_gb": psutil.virtual_memory().total / (1024**3),
            "python_version": sys.version
        },
        "benchmarks": {}
    }
    
    # Add benchmark implementations here
    
    # Save results
    results_file = Path("benchmarks/results") / f"benchmark_{int(time.time())}.json"
    results_file.parent.mkdir(exist_ok=True)
    
    with open(results_file, 'w') as f:
        json.dump(results, f, indent=2)
    
    print(f"Benchmark results saved to {results_file}")

if __name__ == "__main__":
    asyncio.run(run_benchmarks())
