requirements.txt
# CrewAI Framework
crewai==0.70.1
crewai-tools==0.12.1

# LangChain and AI
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16
openai==1.51.0

# Data and Analytics
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
scikit-learn==1.5.1
ta-lib==0.4.32
yfinance==0.2.28
polygon-api-client==1.14.1
alpaca-trade-api==3.2.0

# Web Framework and Dashboard
streamlit==1.38.0
plotly==5.22.0
dash==2.17.1
fastapi==0.112.1
uvicorn==0.30.6
websockets>=10.3,<11.0

# Database and Storage
duckdb==1.0.0
sqlalchemy==2.0.32
redis==5.0.8

# Async and Concurrency
aiohttp==3.10.5
aiofiles==24.1.0
asyncio-mqtt==0.16.2
apscheduler==3.10.4
watchdog==4.0.1

# Configuration and Environment
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
pyyaml>=5.3
toml==0.10.2

# Logging and Monitoring
loguru==0.7.2
prometheus-client==0.20.0

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0

# Utilities
requests==2.32.3
beautifulsoup4==4.12.3
schedule==1.2.2
click==8.1.7
rich==13.7.1
tqdm==4.66.5

# Financial Data Sources
newsapi-python==0.2.7
finnhub-python==2.4.20
quandl==3.7.0

# Cryptocurrency Trading
binance-connector==3.8.1
coinbase-pro==1.1.4
ccxt==4.2.25
pycoingecko==3.1.0
cryptocompare==0.7.5

# Forex Trading
oandapyV20==0.6.3
fxcmpy==1.3.2
alpha-vantage==2.3.1
currencylayer==0.1.12
forexfactory==0.1.0

# Risk Management
riskfolio-lib==6.0.0
# portfolio-optimization==1.1.0  # Package not available, using riskfolio-lib instead

# Machine Learning (Optional)
# tensorflow==2.17.0  # Optional - install separately if needed
# torch==2.4.0  # Optional - install separately if needed
xgboost==2.1.1

# Development Tools
black==24.8.0
flake8==7.1.1
mypy==1.11.2
pre-commit==3.8.0